requirements.txt
Django==4.2.11
djangorestframework==3.15.1
python-dotenv==1.0.1